--- likelion-admin/src/main/java/likelion/univ/alarm/emailsender/EmailSender.java
@@ -1,8 +1,0 @@
-package likelion.univ.alarm.emailsender;
-
-import likelion.univ.email.EmailContent;
-
-public interface EmailSender {
-
-    void send(EmailContent emailContent);
-}
