+++ likelion-infrastructure/src/main/java/likelion/univ/email/sender/EmailContent.java
@@ -1,4 +1,4 @@
-package likelion.univ.email;
+package likelion.univ.email.sender;
 
 import lombok.AccessLevel;
 import lombok.Builder;
